requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Machine learning and clustering
scikit-learn>=1.3.0
scipy>=1.10.0

# Natural language processing (Phase 2)
sentence-transformers>=2.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Utilities
tqdm>=4.65.0
